fix(trainers): cast last_hidden_state to float32 before heads
mDeBERTa-v3 (microsoft/mdeberta-v3-base) emits fp16 tensors from
last_hidden_state on T4 GPU despite no explicit fp16 training args.
Classification / span heads are float32 by default, causing:
  RuntimeError: mat1 and mat2 must have the same dtype, but got Half and Float

Fix: .float() at last_hidden_state level in both trainers so all
heads (cls, start, end in Train_Join; bio head in BiO_Task_mBERT_train)
receive float32 consistently. No-op for mBERT/XLM-R/MuRIL which
already output float32.

+++ Ablations/BiO_Task_mBERT_train.py
@@ -302,7 +302,7 @@ class BIOTagger(torch.nn.Module):
             attention_mask=attention_mask,
             token_type_ids=token_type_ids,
         )
-        seq_out  = self.dropout(outputs.last_hidden_state)  # [B, T, H]
+        seq_out  = self.dropout(outputs.last_hidden_state.float())  # cast: mDeBERTa-v3 can emit fp16 on GPU
         logits   = self.head(seq_out)                        # [B, T, 3]
         return logits
